+++ packages/api-reference/src/components/Content/Models/Models.vue
@@ -1,30 +1,35 @@
 <script setup lang="ts">
 import type { OpenAPIV3_1 } from '@scalar/openapi-types'
 import type { ApiReferenceConfiguration } from '@scalar/types'
+import { computed } from 'vue'
 
 import { Lazy } from '@/components/Lazy'
 import { useNavState } from '@/hooks/useNavState'
+import { getModels } from '@/libs/openapi'
 
 import ClassicLayout from './ClassicLayout.vue'
 import ModernLayout from './ModernLayout.vue'
 
-defineProps<{
+const { document } = defineProps<{
   document: OpenAPIV3_1.Document
   config: ApiReferenceConfiguration
 }>()
 
 const { hash } = useNavState()
+
+/** Returns the model schemas from the document */
+const schemas = computed(() => getModels(document))
 </script>
 <template>
   <Lazy
     id="models"
-    v-if="document?.components?.schemas"
+    v-if="schemas && Object.keys(schemas).length > 0"
     :isLazy="Boolean(hash) && !hash.startsWith('model')">
     <ClassicLayout
       v-if="config?.layout === 'classic'"
-      :schemas="document.components.schemas" />
+      :schemas="schemas" />
     <ModernLayout
       v-else
-      :schemas="document.components.schemas" />
+      :schemas="schemas" />
   </Lazy>
 </template>