+++ packages/api-reference-react/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @scalar/api-reference-react
 
+## 0.7.38
+
+### Patch Changes
+
+- Updated dependencies [83625c6]
+  - @scalar/api-reference@1.34.2
+
+## 0.7.37
+
+### Patch Changes
+
+- Updated dependencies [4d05839]
+  - @scalar/api-reference@1.34.1
+
 ## 0.7.36
 
 ### Patch Changes